requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
python-multipart
pydantic